Improve password modal
